Fix path resolution for child build scripts
When running the top level `build.sh` script from the root of the
repository, paths to the `build.sh` scripts in the immediate
sub-directories are resolved correctly. However, when the top level
build script was run from a different directory, such as the parent
directory of the repo root dir (which is what the DPS system does),
the paths to the child scripts were not resolved, and thus they were not
executed, effectively causing a "no-op" build. These paths are now
correctly resolved no matter what the current directory is when the
top level script is executed.

Fixes #24